requirements.txt
langgraph
langchain-google-genai
langchain-community
yfinance
tavily-python
python-dotenv
pandas
requests
lxml
ipython
